remove unnecessary files, refactor object names
