+++ src/main/scala/com/microsoft/hyperspace/actions/CreateActionBase.scala
@@ -16,10 +16,7 @@
 
 package com.microsoft.hyperspace.actions
 
-import java.io.FileNotFoundException
-
 import org.apache.commons.io.FilenameUtils
-import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.Path
 import org.apache.spark.sql.{DataFrame, SparkSession}
 import org.apache.spark.sql.execution.datasources.{HadoopFsRelation, LogicalRelation, PartitioningAwareFileIndex}
@@ -29,7 +26,6 @@ import org.apache.spark.sql.sources.DataSourceRegister
 
 import com.microsoft.hyperspace.HyperspaceException
 import com.microsoft.hyperspace.index._
-import com.microsoft.hyperspace.index.Content.Directory.FileInfo
 import com.microsoft.hyperspace.index.DataFrameWriterExtensions.Bucketizer
 import com.microsoft.hyperspace.util.{PathUtils, ResolverUtils}
 
@@ -83,7 +79,7 @@ private[actions] abstract class CreateActionBase(dataManager: IndexDataManager) 
                 .Columns(resolvedIndexedColumns, resolvedIncludedColumns),
               IndexLogEntry.schemaString(indexDataFrame.schema),
               numBuckets)),
-          NewContent(absolutePath),
+          Content(absolutePath),
           Source(SparkPlan(sourcePlanProperties)),
           Map())
 
@@ -107,7 +103,7 @@ private[actions] abstract class CreateActionBase(dataManager: IndexDataManager) 
         val files = location.allFiles
         // Note that source files are currently fingerprinted when the optimized plan is
         // fingerprinted by LogicalPlanFingerprint.
-        val sourceDataProperties = Hdfs.Properties(NewContent.fromLeafFiles(files))
+        val sourceDataProperties = Hdfs.Properties(Content.fromLeafFiles(files))
         val fileFormatName = fileFormat match {
           case d: DataSourceRegister => d.shortName
           case other => throw HyperspaceException(s"Unsupported file format: $other")
@@ -129,7 +125,8 @@ private[actions] abstract class CreateActionBase(dataManager: IndexDataManager) 
         IndexConstants.INDEX_NUM_BUCKETS_DEFAULT.toString)
       .toInt
 
-    val (indexDataFrame, resolvedIndexedColumns, _) = prepareIndexDataFrame(spark, df, indexConfig)
+    val (indexDataFrame, resolvedIndexedColumns, _) =
+      prepareIndexDataFrame(spark, df, indexConfig)
 
     // run job
     val repartitionedIndexDataFrame =
@@ -142,22 +139,6 @@ private[actions] abstract class CreateActionBase(dataManager: IndexDataManager) 
         indexDataPath.toString,
         numBuckets,
         resolvedIndexedColumns)
-  }
-
-  private def indexFilesInfo(path: Path): Seq[FileInfo] = {
-    try {
-      val fs = path.getFileSystem(new Configuration)
-      val statuses = fs.listStatus(path)
-
-      // Assuming index directories don't contain nested directories. Only Leaf files.
-      assert(statuses.forall(!_.isDirectory))
-
-      statuses.map(FileInfo(_)).toSeq
-    } catch {
-      // FileNotFoundException is an expected exception before index gets created.
-      case _: FileNotFoundException => Seq()
-      case e: Throwable => throw e
-    }
   }
 
   private def resolveConfig(
