+++ src/test/scala/com/microsoft/hyperspace/index/E2EHyperspaceRulesTests.scala
@@ -435,7 +435,7 @@ class E2EHyperspaceRulesTests extends HyperspaceSuite with SQLHelper {
   }
 
   private def getIndexFilesPath(indexName: String): Seq[Path] = {
-    NewContent(new Path(
+    Content(new Path(
       systemPath,
       s"$indexName/${IndexConstants.INDEX_VERSION_DIRECTORY_PREFIX}=0")).files
   }
